% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_anoms.R
\name{calc_anoms}
\alias{calc_anoms}
\title{Calculate the anomaly values for a table produced by gcm_anomalies()}
\usage{
calc_anoms(datin)
}
\arguments{
\item{datin}{a \code{tbl_df} produced by gcm_anomalies().}
}
\description{
Calculate the anomaly values for a table produced by gcm_anomalies()
}
