% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataEliminator.R
\name{dataEliminator}
\alias{dataEliminator}
\title{Eliminate daily values from a monthly data set}
\usage{
dataEliminator(month, csv, NAs, sampling = "z", variables = c("max", "min",
  "mean"), simplify = FALSE, interpolate = FALSE)
}
\arguments{
\item{month}{data.frame; a data.frame with a single year-month of data with no missing values}

\item{csv}{character; path to a .csv file containing a single year-month of data with no missing values}

\item{NAs}{numeric; a vector of the number of NA values to test}

\item{sampling}{character; the type of sampling to use: (r)andom, (c)onsecutive}

\item{variables}{character; the names of the variables to test (we will try to auto-idenify the column number)}

\item{simplify}{Boolean; whether to return simplified results}

\item{interpolate}{Boolean; whether to use linear interpolation to approximate the missing values}
}
\description{
Eliminate daily values from a monthly data set
}
