% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataEliminatorMassive.R
\name{dataEliminatorMassive}
\alias{dataEliminatorMassive}
\title{Perform the mass elimination of daily values from a monthly data set}
\usage{
dataEliminatorMassive(numberTests, month, csv, NAs, sampling = "z",
  variable = c("max", "min", "mean"), verbose = FALSE,
  interpolate = FALSE)
}
\arguments{
\item{numberTests}{integer; the number of times to repeat the test}

\item{month}{data.frame; a data.frame with a single year-month of data with no missing values}

\item{csv}{character; path to a .csv file containing a single year-month of data with no missing values}

\item{NAs}{numeric; a vector of the number of NA values to test}

\item{sampling}{character; the type of sampling to use: (r)andom, (c)onsecutive}

\item{variable}{character; the name of the variable to test (we will try to auto-idenify the column number)}

\item{verbose}{Boolean; whether the function should be verbose}

\item{interpolate}{Boolean; whether to use linear interpolation to approximate the missing values}
}
\description{
Perform the mass elimination of daily values from a monthly data set
}
