% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataEliminatorThorough.R
\name{dataEliminatorThorough}
\alias{dataEliminatorThorough}
\title{Perform the mass elimination of consecutive daily values from a monthly data set}
\usage{
dataEliminatorThorough(month, csv, NAs, variables = c("max", "min", "mean"),
  simplify = FALSE, interpolate = FALSE)
}
\arguments{
\item{month}{data.frame; a data.frame with a single year-month of data with no missing values}

\item{csv}{character; path to a .csv file containing a single year-month of data with no missing values}

\item{NAs}{numeric; a vector of the number of NA values to test}

\item{variables}{character; the names of the variables to test (we will try to auto-idenify the column number)}

\item{simplify}{Boolean; whether to return simplified results}

\item{interpolate}{Boolean; whether to use linear interpolation to approximate the missing values}
}
\description{
Perform the mass elimination of consecutive daily values from a monthly data set
}
