% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaDTD.R
\name{deltaDTD}
\alias{deltaDTD}
\title{Calculate deltaDTD on any time scale}
\usage{
deltaDTD(data, period = "z", QA = TRUE)
}
\arguments{
\item{data}{data.frame; define the data that should be analyzed}

\item{period}{character; The period for the output data. One of "daily", "monthly", "seasonal", or "annual".}

\item{QA}{logical; whether to filter those values that are missing 20\% of observations}
}
\description{
"A function that calculates DTD, SD, G, and deltaDTD on daily, monthly, seasonal, or annual time scales. It automatically eliminates aggregate values with higher than 20\% missing values. This function requres a data frame with dates in column 1 called 'Date', Tmax in column 2 called 'MaxTemp', Tmin in column 3 called 'MinTemp'. No other data is necessary Note that the data aquired though the \code{canadaHCD} package meets these needs."
}
\examples{
\dontrun{deltaDTD(tor_dly, "annual")}
}
\author{
Conor I. Anderson
}

