% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_snaps.R
\name{describe_snaps}
\alias{describe_snaps}
\title{Describe run-length-encoded events}
\usage{
describe_snaps(run_lengths, threshold = 2, vec, f = NULL, ...)
}
\arguments{
\item{run_lengths}{an object of class \code{rle}, of some criteria for an "event"}

\item{threshold}{the minimum length of the events to consider (default: 2)}

\item{vec}{(optional) a vector of data for which \code{f} should be applied}

\item{f}{(optional) a function to apply to the values of \code{vec} where events are \code{TRUE}.}

\item{...}{additional parameters to be passed to \code{f}}
}
\value{
a vector
}
\description{
Given an object of class \code{rle}, prints a vector of the length of
\code{sum(run_lengths$lengths)}, with each length repeated. It is useful to
annotate continuous events in a data frame.
}
\details{
If passed a function along with an additional vector. The function is applied
to the values of the vector when each event is \code{TRUE}.
}
\examples{
# Create some "weather"
weather <- rnorm(1000)
# Get the lengths of "cold" events where the temperature is below zero for more than 3 days
describe_snaps(rle(weather < 0), 3)
}
