% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_means.R
\name{ensemble_means}
\alias{ensemble_means}
\title{Calculate the average of all ensembles for the same Variable/Model/Scenario}
\usage{
ensemble_means(datin)
}
\arguments{
\item{datin}{a \code{tbl_df} produced by gcm_anomalies().}
}
\description{
Calculate the average of all ensembles for the same Variable/Model/Scenario
}
