% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcm_anomalies.R
\name{gcm_anomalies}
\alias{gcm_anomalies}
\title{Generate table GCM anomaly data}
\usage{
gcm_anomalies(dir = getwd(), lat, lon, timeslice = 0, baseline,
  proj = list(2011:2040, 2041:2070, 2071:2100), simple_names = FALSE,
  calc_anom = TRUE, ensemble_average = TRUE)
}
\arguments{
\item{dir}{character; path to the directory containing the \code{.nc} files you wish to process}

\item{lat}{numeric; the latitude of the target cell}

\item{lon}{numeric; the longitude of the target cell}

\item{timeslice}{numeric; 1-16: annual (0), monthly (1-12), seasonal (13-16)}

\item{baseline}{the baseline on which the check the data (defaults to 1971:2000)}

\item{proj}{list containing vectors of the projection periods (defaults to 2020s, 2050s, and 2080s)}

\item{simple_names}{Boolean; whether to simplify the projection interval names (ie. 2020s instead of 2011-2040)}

\item{calc_anom}{Boolean; whether to calculate anomaly values.}

\item{ensemble_average}{Boolean; whether to average different ensembles for the same variable/model/scenario.}
}
\value{
a class \code{data.frame}
}
\description{
Generate table GCM anomaly data
}
\examples{
\dontrun{annual <- gcm_anomalies(getwd(), -7.023833, -76.465222, 0, 1971:2000, calc_anoms = TRUE)}
}
