% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wkt_csv.R
\name{generate_wkt_csv}
\alias{generate_wkt_csv}
\title{Generate WKT-formatted polygons for gridded parsed ASCII data}
\usage{
generate_wkt_csv(datain, name, filename, calc_mean = TRUE,
  remove_missing = TRUE)
}
\arguments{
\item{datain}{object; a matrix of parsed gridded ASCII data generated using \code{\link{parse_ASCII_grid}}}

\item{name}{character; the name of the matrix (or matrices) to generate \code{.csv} files for, e.g. \code{"Dec 2013"}}

\item{filename}{character; the name to save the file under}

\item{calc_mean}{Boolean; if the vector of names has \code{length(name) > 1}, the we will calculate the average for the selected matrices}

\item{remove_missing}{Boolean; whether we should remove missing values in the mean calculation}
}
\description{
This function creates a \code{.csv} file from parsed gridded ASCII data, with a properly formatted WKT polygon definition for each grid.
}
\examples{
# Generate an anomaly maps for Winter 2013-14
\dontrun{generate_wkt_csv(dat, c("Dec 2013", "Jan 2014", "Feb 2014"))}
}
