% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcm_files.R
\name{get_gcm_files}
\alias{get_gcm_files}
\title{Generate a list of GCM files to pass to \code{gcm_anomalies()}}
\usage{
get_gcm_files(dir, baseline = NULL, projection = NULL, scenario = NULL,
  ensemble = NULL)
}
\arguments{
\item{dir}{character; the directory containing the files.}

\item{baseline}{numeric; a vector years that should be covered by history (\code{NULL} to ignore)}

\item{projection}{numeric; a vector years that should be covered by the models (\code{NULL} to ignore)}

\item{scenario}{character; the scenario to filter on (optional)}

\item{ensemble}{character; the ensemble to fulter on (optional)}
}
\value{
a list of filenames
}
\description{
Generate a list of GCM files to pass to \code{gcm_anomalies()}
}
\examples{
\dontrun{get_gcm_files(getwd(), 1971:2000, scenario = c("rcp85", "rcp45", "rcp26", "rcp60"), ensemble = "r1i1p1")}
}
