% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_value_lab.R
\name{missing_data_lab}
\alias{missing_data_lab}
\title{Eliminate daily values within months of a data set and report the results}
\usage{
missing_data_lab(monthin, NAs, sampling = "z", no_of_tests = 1,
  variables = c("max", "min", "mean"), cores)
}
\arguments{
\item{monthin}{data.frame; a data.frame with a single year-month of data with no missing values}

\item{NAs}{numeric; a vector of the number of NA values to test}

\item{sampling}{character; the type of sampling to use: (r)andom, (c)onsecutive}

\item{no_of_tests}{integer; the number of tests to run; for consecutive sampling, tops at N - k + 1}

\item{variables}{character; the names of the variables to test (we will try to auto-idenify the column number)}

\item{cores}{numeric; the number of cores to parallize over. Defaults to \code{parallel::detectCores() - 1}}
}
\description{
Eliminate daily values within months of a data set and report the results
}
