% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_daily.R
\name{solar_daily}
\alias{solar_daily}
\title{Daily extremes using the Night and Day Climatological Observing Window}
\source{
\insertRef{zaknic-catovic_2018_comparison}{claut}
}
\usage{
solar_daily(dat, lat, lon, tz, stimes = FALSE, ...)
}
\arguments{
\item{dat}{data frame; the data to analyze; the first column must be date times; subsequent columns should contain for which to detect extremes}

\item{lat}{numeric; the latitude of the observer}

\item{lon}{numeric; the longitude of the observer}

\item{tz}{integer; the number of hours shifted from GMT, e.g. for GMT-5, pass \code{-5}}

\item{stimes}{Boolean; whether the sunrise and sunset should be included in the results table}

\item{...}{additional parameters to pass to \code{\link{max}} and \code{\link{min}}, e.g. \code{na.rm = TRUE}}
}
\value{
A table of class \code{\link{tbl_df}}.
}
\description{
Calculates daily maxima and minima after \insertCite{zaknic-catovic_2018_comparison}{claut}
}
\author{
Conor I. Anderson
}
