% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subensemble.R
\name{subensemble}
\alias{subensemble}
\title{Find optimal sub-ensembles}
\usage{
subensemble(k, datain, target, stdev = NULL, model_names = TRUE,
  single_result = FALSE, parallel = FALSE,
  no_cores = parallel::detectCores() - 1)
}
\arguments{
\item{k}{integer; the sub-ensemble size(s) to test}

\item{datain}{numeric; a vector or data.frame containing the GCM anomalies for each baseline series}

\item{target}{numeric; a vector of the baseline averaegs to test against}

\item{stdev}{numeric; the standard deviation of the annual average baseline value to test against}

\item{model_names}{Boolean; whether the datain includes a column of model names}

\item{single_result}{Boolean; whether to return one optimal sub-ensemble size}

\item{parallel}{Boolean; whether to perform operations in parallel}

\item{no_cores}{integer; the numer of CPU cores to use for parallel processing}
}
\value{
a data.frame with one row for each sub-ensemble size
}
\description{
Find optimal sub-ensembles
}
