% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suntimes.R
\name{suntimes}
\alias{suntimes}
\title{Sunrise, solar noon, and sunset times}
\source{
\insertRef{noaa_esrl_solar}{claut}
}
\usage{
suntimes(lat, lon, tz, date, out = c("sunrise", "solar_noon", "sunset"))
}
\arguments{
\item{lat}{numeric; the latitude of the observer}

\item{lon}{numeric; the longitude of the observer}

\item{tz}{integer; the number of hours shifted from GMT, e.g. for GMT-5, pass \code{-5}}

\item{date}{date; the date to get the sun times for}

\item{out}{character; one of "sunrise", "solar_noon", or "sunset". Leaving this parameter blank will retrun a list of all three.}
}
\value{
POSIXct date time for selected sun time
}
\description{
Calculations are based on NOAA's solar calculator, which is, in turn, based on Astronomical Algorithms, by Jean Meeus.
}
\examples{
## Get today's sunrise time for Toronto Pearson Airport
suntimes(lat = 43.68, lon = -79.63, tz = -5, date = Sys.Date())
}
