% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaDTD.R
\name{deltaDTD}
\alias{deltaDTD}
\title{Calculate deltaDTD on any time scale}
\usage{
deltaDTD(datain, period = "z", max_NA = 0.2)
}
\arguments{
\item{datain}{datain.frame; define the datain that should be analyzed}

\item{period}{character; The period for the output datain. One of "daily", "monthly", "seasonal", or "annual".}

\item{max_NA}{numeric; The maximum proportion of missing data to allow (e.g. 0.2 for 20\% missing values)}
}
\description{
A function that calculates DTD, SD, G, and deltaDTD on daily, monthly, seasonal, or annual time scales. It automatically eliminates aggregate values with higher than 20\% missing values. This function requres a datain frame with dates in column 1 called 'Date', Tmax in column 2 called 'MaxTemp', Tmin in column 3 called 'MinTemp'. No other datain is necessary Note that the datain aquired though the \code{canadaHCD} package meets these needs.
}
\examples{
\dontrun{deltaDTD(tor_dly, "annual")}
}
\author{
Conor I. Anderson
}
