% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcm_anomalies.R
\name{gcm_anomalies}
\alias{gcm_anomalies}
\title{Generate table GCM anomaly data}
\usage{
gcm_anomalies(dir = getwd(), files, lat, lon, timeslice = 0, baseline,
  start = 2011, simple_names = FALSE, calc_anom = TRUE,
  ensemble_average = TRUE, cache_dir = NULL)
}
\arguments{
\item{dir}{character; path to the directory containing the \code{.nc} files you wish to process}

\item{files}{character; list of filenames to process. Any list item of >1 will be processed sequentially.}

\item{lat}{numeric; the latitude of the target cell}

\item{lon}{numeric; the longitude of the target cell}

\item{timeslice}{numeric; 1-16: annual (0), monthly (1-12), seasonal (13-16)}

\item{baseline}{the baseline on which the check the data (defaults to 1971:2000)}

\item{start}{projection start year (defaults to 2011)}

\item{simple_names}{Boolean; whether to simplify the projection interval names (ie. 2020s instead of 2011-2040)}

\item{calc_anom}{Boolean; whether to calculate anomaly values.}

\item{ensemble_average}{Boolean; whether to average different ensembles for the same variable/model/scenario.}

\item{cache_dir}{character; a directory to cache results (defaults to NULL, i.e. no cache.)}
}
\value{
a class \code{data.frame}
}
\description{
Generate table GCM anomaly data
}
\examples{
\dontrun{annual <- gcm_anomalies(getwd(), -7.023833, -76.465222, 0, 1971:2000, calc_anom = TRUE)}
}
