% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimData.R
\name{trimData}
\alias{trimData}
\title{Trim daily data to a desired start and end year}
\usage{
trimData(data, start, end)
}
\arguments{
\item{data}{data.frame; define the data that should be trimmed}

\item{start}{numerical; the first year to include}

\item{end}{numerical; the last year to include}
}
\description{
Trim a daily data set to a desired start and end year. Originally written for data obtained using the \code{canadaHCD} package.
}
\examples{
\dontrun{trimData(tor_dly, 1991, 2010)}
}
\author{
Conor I. Anderson
}

